==> analysis/side_events/0_get_side_events_archive.ipynb <==
import requests
from bs4 import BeautifulSoup as bs
import pandas as pd
import numpy as np
from selenium import webdriver
from selenium.webdriver.support.ui import Select
import time

## Extract COP archive side events data


url = 'https://seors.unfccc.int/applications/seors/reports/archive.html'
r = requests.get(url).content.decode()
soup = bs(r, 'html')

# Compile the historical events
items = soup.select('[name=session_id] option[value]')
event_table = pd.DataFrame(data={"code":np.nan,"event":np.nan},index=np.arange(len(items))[1:])
for i in np.arange(len(items))[1:]:
    this_item = str(items[i])
    event_table.loc[i,"code"] = this_item.split("=")[1].split(">",1)[0][1:-1]
    event_table.loc[i,"event"] = this_item.split(">")[1][:-8]
event_table_cop = event_table[event_table["code"].str.contains("COP")]
event_table_cop = event_table_cop[event_table_cop["code"] != "COP 23"]
event_table_cop = event_table_cop.set_index("code").reset_index()
for i in event_table_cop.index:
    event_table_cop.loc[i,"name"] = "COP"+ str(28-i).zfill(2)
event_table_cop.set_index("code").to_csv("COP_archive_side_events_2024.csv")
event_table_cop

### Extract data 

# set up chrome driver
# see start of the file on how to set this up 
options = webdriver.ChromeOptions()
options.add_argument('--ignore-certificate-errors')
options.add_argument('--incognito')
options.add_argument('--headless') # make it work without opening a window 
driver = webdriver.Chrome("C:\Program Files\Java\chromedriver-win64\chromedriver.exe", options=options) # USER DEFINED FILEPATH

for i in event_table_cop.index:
    event_code = event_table_cop.loc[i,"code"]
    event_name = event_table_cop.loc[i,"name"]

    # select the category at the drop down 
    driver.get(url)
    select = Select(driver.find_element_by_name('session_id')) # select category by the drop down 
    select.select_by_value(event_code) # select by value
    time.sleep(1)
    driver.find_element_by_name('Submit').submit() # click the "refresh" button 
    time.sleep(1)

    page_source = driver.page_source
    soup_events = bs(page_source, 'html')

    table_events = soup_events.find_all('table')[4]
    df_events = pd.read_html(str(table_events))[0]
    df_events = df_events[df_events["Date"].isna()==False]
    df_events = df_events[df_events["Organizer"].isna()==False] # some attachments are shifted weird
    df_events = df_events.reset_index()
    df_events = df_events.drop(columns={"Attachments","index"})
    df_events = df_events.drop_duplicates()
    df_events = df_events.drop
    df_events.index.names = ['Code']
    df_events = df_events.reset_index()
    df_events["SE_Code"] = event_name+ "-" + df_events["Code"].astype(str).str.zfill(3)
    
    if i == event_table_cop.index[0]:
        df_events_all = df_events
    else:
        df_events_all = pd.concat([df_events_all,df_events])
        
df_events_all.set_index("SE_Code").to_csv("COP_archive_side_events_compiled_raw_2024.csv")
df_events_all

df_events
