==> analysis/side_events/1_clean_data.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import re # recognize phone number

from langdetect import detect, detect_langs
from deep_translator import GoogleTranslator

### Clean data for archived side events

df_events_all = pd.read_csv("COP_archive_side_events_compiled_raw_2024.csv")

# split up detailed titles, desciprions, detailed organizer information
df_events_all[['Title/Description','Speakers']] = df_events_all["Title/theme/speakers"].str.split('  ', expand=True)
df_events_all = df_events_all[df_events_all['Title/Description']!="test test"] ## REMOVE test events
df_events_all = df_events_all.set_index("SE_Code").reset_index()
df_events_all["Organizer_Affiliation"] = np.nan
df_events_all["Organizer_Affiliation"] = df_events_all["Organizer_Affiliation"].astype(object)
df_events_all["Organizer_Name"] = np.nan
df_events_all["Organizer_Name"] = df_events_all["Organizer_Name"].astype(object)

# set up edges dataframe
edges_side_events_archive = pd.DataFrame(data={"Title/Description":np.nan,"SE_Code":np.nan},index=[0])

edges_count = 0
for i in df_events_all.index:
    orgnzr = df_events_all["Organizer"][i]
    org_list = []
    organizer_list = []
    for unit in orgnzr.split("  "):
        if "@" in unit:
            org_name = unit.rsplit("@",1)[0]
            org_name = org_name.rsplit(" ",1)[0]
            org_list.append(org_name)
        elif "+" not in unit:
            organizer_list.append(unit)
    df_events_all.at[i,"Organizer_Affiliation"] = list(set(org_list)) # delete duplicates
    df_events_all.at[i,"Organizer_Name"] = organizer_list
    df_events_all.loc[i,"Organizer_Count"] = len(org_list)
    
    org_list_unique = list(set(org_list))
    for n in np.arange(len(org_list_unique))[:-1]:
        source = org_list_unique[n]
        for m in np.arange(len(org_list_unique))[n+1:]:
            edges_side_events_archive.loc[edges_count,"Source_Name"] = source
            edges_side_events_archive.loc[edges_count,"Target_Name"] = org_list_unique[m]
            edges_side_events_archive.loc[edges_count,"Title/Description"] = df_events_all.loc[i,"Title/Description"]
            edges_side_events_archive.loc[edges_count,"SE_Code"] = df_events_all.loc[i,"SE_Code"]
            edges_side_events_archive.loc[edges_count,"Weight"] = 1/(np.arange(len(org_list_unique)).sum())
            edges_count = edges_count + 1
            
df_events_all["Year"] = df_events_all["Date"].str[-4:].astype(int)
df_events_all.set_index("SE_Code").to_csv("all_side_events.csv")
edges_side_events_archive.set_index("SE_Code").to_csv("Edges_All.csv")
df_events_all

### clean the dataset
- remove incomplete descriptions
- remove the word Cancelled 

# filter those with "TBC" in title/description, excluding "(TBC)" which only indicates unknown speakers in the description
TBC_df = df_events_all[df_events_all["Title/Description"].str.contains("TBC")].copy()
TBC_df = TBC_df[TBC_df["Title/Description"].str.contains("\(TBC\)")==False]
df_events_complete = df_events_all[df_events_all.index.isin(TBC_df.index)==False]
#(* tbc)
tbc_df2 = df_events_all[df_events_all["Title/Description"].str.contains("tbc")].copy()
tbc_df2 = tbc_df2[tbc_df2["Title/Description"].str.contains("\(* tbc\)")==False]
df_events_complete = df_events_complete[df_events_complete.index.isin(tbc_df2.index)==False]

# remove those with "TBA", "tbc"
strings_to_remove = ["TBA", "Tbc", "WILPF Italy WILPF Italy", "President Evo Morales meets IPOs",
                    "Coming soon ..", "COP Presidency Gender Event Presidency Gender Event"]
for special_string in strings_to_remove:
    df_events_complete = df_events_complete[df_events_complete["Title/Description"].str.contains(special_string)==False] 
    
# remove descriptions shorter than 10 words
df_events_complete["word_count"] = df_events_complete['Title/Description'].str.split().str.len()
df_events_complete = df_events_complete[df_events_complete["word_count"]>10]

df_events_all[df_events_all.index.isin(df_events_complete.index)==False].set_index("SE_Code").to_excel("note_side_events_incomplete.xlsx")
len(df_events_complete)

# cancelled events
df_events_complete["Title/Description"] = df_events_complete["Title/Description"].str.replace("Cancelled","")

df_events_complete.to_csv("all_side_events_complete.csv",index=False)

### identify languages

df_events_complete_lang = df_events_complete.copy()
for i in df_events_complete_lang.index:
    text = df_events_complete_lang.loc[i,"Title/Description"]
    lang = detect(text)
    df_events_complete_lang.loc[i,"lang_most"] = lang
    if lang == "en":
        df_events_complete_lang.loc[i,"Title/Description_new"] = text
    else:
        df_events_complete_lang.loc[i,"Title/Description_new"] = GoogleTranslator(source=lang, target='en').translate(text)
    if i%1000==0:
        print("Completed detection on for "+str(i)+ " entries")
        print(detect(text))
df_events_complete_lang[df_events_complete_lang["lang_most"]!="en"].set_index("SE_Code").to_excel("note_side_events_language.xlsx")
df_events_complete_lang.to_csv("all_side_events_complete_lang.csv",index=False)
df_events_complete_lang

# split up detailed titles, desciprions, detailed organizer information
df_events_all[['Title/Description','Speakers']] = df_events_all["Title/theme/speakers"].str.split('  ', expand=True)
df_events_all = df_events_all.set_index("SE_Code").reset_index()
df_events_all["Organizer_Affiliation"] = np.nan
df_events_all["Organizer_Affiliation"] = df_events_all["Organizer_Affiliation"].astype(object)
df_events_all["Organizer_Name"] = np.nan
df_events_all["Organizer_Name"] = df_events_all["Organizer_Name"].astype(object)

# set up edges dataframe
edges_side_events_archive = pd.DataFrame(data={"Title/Description":np.nan,"SE_Code":np.nan},index=[0])

edges_count = 0
for i in df_events_all.index:
    orgnzr = df_events_all["Organizer"][i]
    org_list = []
    organizer_list = []
    for unit in orgnzr.split("  "):
        if "@" in unit:
            org_name = unit.rsplit("@",1)[0]
            org_name = org_name.rsplit(" ",1)[0]
            org_list.append(org_name)
        elif "+" not in unit:
            organizer_list.append(unit)
    df_events_all.at[i,"Organizer_Affiliation"] = list(set(org_list)) # delete duplicates
    df_events_all.at[i,"Organizer_Name"] = organizer_list
    df_events_all.loc[i,"Organizer_Count"] = len(org_list)
    
    org_list_unique = list(set(org_list))
    for n in np.arange(len(org_list_unique))[:-1]:
        source = org_list_unique[n]
        for m in np.arange(len(org_list_unique))[n+1:]:
            edges_side_events_archive.loc[edges_count,"Source_Name"] = source
            edges_side_events_archive.loc[edges_count,"Target_Name"] = org_list_unique[m]
            edges_side_events_archive.loc[edges_count,"Title/Description"] = df_events_all.loc[i,"Title/Description"]
            edges_side_events_archive.loc[edges_count,"SE_Code"] = df_events_all.loc[i,"SE_Code"]
            edges_side_events_archive.loc[edges_count,"Weight"] = 1/(np.arange(len(org_list_unique)).sum())
            edges_count = edges_count + 1
    
df_events_all.set_index("SE_Code").to_csv("COParchive_side_events_compiled.csv")
edges_side_events_archive.set_index("SE_Code").to_csv("Edges_Archive.csv")
