==> analysis/side_events/3_isolate_topic.ipynb <==
import pandas as pd
import numpy as np

all_edges = pd.read_excel("Edges_All_withNodes.xlsx")
all_nodes = pd.read_excel("Nodes_All.xlsx")

### energy 

energy_events = pd.read_csv("extracted_events_energy.csv")
all_edges_energy = all_edges[all_edges["SE_Code"].isin(energy_events["SE_Code"])]
all_edges_energy = pd.merge(all_edges_energy,energy_events[["SE_Code","new_topics_ctfidf"]],on="SE_Code") # add topics
all_edges_energy.to_excel("Edges_energy.xlsx")
all_edges_energy.head()

energy_orgs = pd.read_excel("affiliation/organiser_weight_energy.xlsx")
energy_orgs = pd.merge(energy_orgs,all_nodes,left_on="clean_org",right_on="Name",how="left")
individual_orgs = energy_orgs[energy_orgs["ID"].isna()]
count = 0
for i in individual_orgs.index:
    energy_orgs.loc[i,"ID"] = "ind-"+str(count)
    count = count + 1
energy_orgs.to_excel("Nodes_energy.xlsx",index=False)
energy_orgs.head()

### food

food_events = pd.read_csv("extracted_events_food.csv")
all_edges_food = all_edges[all_edges["SE_Code"].isin(food_events["SE_Code"])]
all_edges_food = pd.merge(all_edges_food,food_events[["SE_Code","new_topics_ctfidf"]],on="SE_Code") # add topics
all_edges_food.to_excel("Edges_food.xlsx")
all_edges_food
