==> analysis/side_events/4_find_nodes_interest_groups.ipynb <==
import pandas as pd
import numpy as np

### oil and gas

delegates_fossil = pd.read_excel("../delegates/COP28_delegates_ff_organizations_renamed_filtered_check.xlsx")
energy_nominator_rename = pd.read_excel("nominator_energy_match.xlsx")
energy_sponsor = delegates_fossil[["nominator"]].drop_duplicates()
for i in energy_nominator_rename.index:
    energy_sponsor["nominator"] = energy_sponsor["nominator"].replace(energy_nominator_rename.loc[i,"original_name"],energy_nominator_rename.loc[i,"renamed"])
energy_sponsor = energy_sponsor.drop_duplicates()
energy_sponsor

energy_nodes = pd.read_excel("Nodes_energy.xlsx")
energy_nodes["ff_nominator"] = 0
matched_nominator = energy_nodes[energy_nodes["clean_org"].isin(energy_sponsor["nominator"])]
energy_nodes.loc[matched_nominator.index,"ff_nominator"] = 1
energy_nodes.to_excel("Nodes_energy_nominator.xlsx",index=False)
energy_nodes

### meat and dairy

delegates_food = pd.read_excel("../delegates/COP28_delegates_meat&dairy_renamed_filtered_checked.xlsx")
food_sponsor = delegates_food[delegates_food["type"]=='Non-governmental organizations'][["nominator"]].drop_duplicates()
food_sponsor

food_edges = pd.read_excel("edges/Edges_food.xlsx")
meat_dairy_list = ["meat","dairy","livestock","poultry"]

meat_dairy_edges = edge_search(meat_dairy_list, food_edges, "Title/Description") 
food_unique = set(list(food_edges["Source"].unique()) + list(food_edges["Target"].unique()))
meat_dairy_unique = set(list(meat_dairy_edges["Source"].unique()) + list(meat_dairy_edges["Target"].unique()))

food_nodes = all_nodes[all_nodes["ID"].isin(food_unique)]
food_nodes["short_name"] = food_nodes["Name"].str.split(" \(",expand=True)[0]

# nodes on meat and dairy events
food_nodes.loc[list(meat_dairy_unique),"meat_event"] = 1
food_nodes["meat_event"] = food_nodes["meat_event"].fillna(0)

# nodes that nominated meat companies
meat_nominator = food_nodes[food_nodes["short_name"].isin(food_sponsor["nominator"].unique())]
food_nodes.loc[meat_nominator.index,"meat_nominator"] = 1
food_nodes["meat_nominator"] = food_nodes["meat_nominator"].fillna(0)
food_nodes.to_excel("nodes/Nodes_food.xlsx", index=False)
food_nodes
